\name{profX}
\docType{data}
\alias{profX}
\title{A matrix with 26 language - profiles}
\description{
This is a 26 * 11.022 matrix. Each line represents one language, each column
one n-gram. The values reveal, how often a specific n-gram has been counted
in the trainingdata - if it did belong to the top 1000.}
\usage{data("profX")}
\format{A 26 * 11.022 matrix}
\source{ECI, Gutenberg, private}
