\name{profX_ger_eng}
\docType{data}
\alias{profX_ger_eng}
\title{A matrix with 2 language - profiles}
\description{
This is a matrix. Each line represents one language, each column
one n-gram. The values reveal, how often a specific n-gram has been counted
in the trainingdata - if it did belong to the top 1000.}
\usage{data("profX_ger_eng")}
\format{A matrix}
\source{ECI, Gutenberg, private, The Economist}
